REINTEGRATED branch 'branches/distribution_rework'. Introducing a new grid distribution implementation, which is capable of distributing / redistributing adaptive multigrid hierarchies. Distributed grids can be redistributed with arbitrary partition maps. Old distribution code has been removed, to clean up the code-base. While the methods have been tested, errors may still occur in some applications. The rework made the adjustment of some scripts necessary. I hope that everything works as one would expect.

+++ ugbase/common/profiler/profile_node.cpp
@@ -434,11 +434,11 @@ void WriteProfileData(const char *filename)
 		if(bProfileAll)
 		{
 			std::vector<ug::BinaryBuffer> buffers(pcl::GetNumProcesses()-1);
-			for(size_t i=1; i<pcl::GetNumProcesses(); i++)
+			for(int i=1; i<pcl::GetNumProcesses(); i++)
 				ic.receive_raw(i, buffers[i-1]);
 			ic.communicate();
 
-			for(size_t i=1; i<pcl::GetNumProcesses(); i++)
+			for(int i=1; i<pcl::GetNumProcesses(); i++)
 			{
 				f << "\n<core id=\"" << i << "\">";
 				string s;
